requirements.txt
fastapi
uvicorn[standard]
pydantic
pymongo
pydantic_settings
jwt
selenium
webdriver-manager
pandas
openpyxl
unidecode
PyMuPDF
Pillow
pytesseract
python-multipart